Split some components to a different file for easy reading
